Move error setup to bootstrap.
